+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 26.03.1+ (???)
 ------------------------------------------------------------------------
+- Feature: [#3639] The Locomotion title screen music can now be listened to during scenario play.
 - Fix: [#3334] Auto order of cars with centrePosition flag incorrectly calculated.
 - Fix: [#3638] Loan can go negative.